+++ storage/aptosdb/src/db/include/aptosdb_testonly.rs
@@ -16,7 +16,8 @@ impl AptosDB {
             false,
             BUFFERED_STATE_TARGET_ITEMS,
             DEFAULT_MAX_NUM_NODES_PER_LRU_CACHE_SHARD,
-            false,
+            false, /* indexer */
+            false, /* indexer async v2 */
         )
     }
 
@@ -34,16 +35,17 @@ impl AptosDB {
             false,
             NO_OP_STORAGE_PRUNER_CONFIG, /* pruner */
             db_config,
-            false,
+            false, /* indexer */
             BUFFERED_STATE_TARGET_ITEMS,
             max_node_cache,
+            false, /* indexer async v2 */
         )
         .expect("Unable to open AptosDB")
     }
 
     /// This opens db in non-readonly mode, without the pruner and cache.
     pub fn new_for_test_no_cache<P: AsRef<Path> + Clone>(db_root_path: P) -> Self {
-        Self::new_without_pruner(db_root_path, false, BUFFERED_STATE_TARGET_ITEMS, 0, false)
+        Self::new_without_pruner(db_root_path, false, BUFFERED_STATE_TARGET_ITEMS, 0, false, false)
     }
 
     /// This opens db in non-readonly mode, without the pruner, and with the indexer
@@ -53,7 +55,8 @@ impl AptosDB {
             false,
             BUFFERED_STATE_TARGET_ITEMS,
             DEFAULT_MAX_NUM_NODES_PER_LRU_CACHE_SHARD,
-            true,
+            true, /* indexer */
+            true, /* indexer async v2 */
         )
     }
 
@@ -67,7 +70,8 @@ impl AptosDB {
             false,
             buffered_state_target_items,
             DEFAULT_MAX_NUM_NODES_PER_LRU_CACHE_SHARD,
-            false,
+            false, /* indexer */
+            false, /* indexer async v2 */
         )
     }
 
@@ -78,7 +82,8 @@ impl AptosDB {
             true,
             BUFFERED_STATE_TARGET_ITEMS,
             DEFAULT_MAX_NUM_NODES_PER_LRU_CACHE_SHARD,
-            false,
+            false, /* indexer */
+            false, /* indexer async v2 */
         )
     }
